Add files via upload
The idea of our project We have designed a program in
Java that connects to databases. We have designed a train station that helps the user to book a trip for him or her by creating an account. We also did not design our program only for users who wanted to book at the station. We added interfaces to the Admin and finally, the employees can Perform some
Different operations. Our legislators contain 12 classes